requirements.txt
telethon==1.34.0
python-decouple==3.8
